* fix for Kaffe_JNI_Wrapper (take translator lock) * introduced staticLockIsInitialized() macro * implemented call and branch target alignment (experimental)

+++ kaffe/kaffevm/jit/icode.c
@@ -3280,7 +3280,7 @@ softcall_lookupmethod(SlotInfo* dst, Method* meth, SlotInfo* obj)
 	/* 'obj' must be written back since it will be reused */
 	prepare_function_call();
 	/* FIXME: should there be a pusharg_int_const? */
-	pusharg_ref_const((void*)meth->idx, 2);		
+	pusharg_ref_const((void*)(int)meth->idx, 2);		
 	pusharg_class_const(meth->class, 1);
 	pusharg_ref(obj, 0);
 	call_soft(soft_lookupmethod);
